Big ol' redesign: distinguish realtime anomalies.
It used to be that, say, a G1c anomaly depended on external-graphs: if
you included realtime, G1c might use realtime edges. This was consistent
with Adya, but was misleading if you didn't understand this implicit
assumption, and led to incorrect statements about what consistency
models were disallowed.

We now report these anomalies separately, so G1c is *purely* wr and ww
edges, and G1c-realtime may also include realtime dependencies.
Likewise, we have G0-process, G1c-process, G-single-process, etc.
Honestly, it's a LOT of anomalies, but a.) I don't think we're going to
add anything else BUT realtime and process, and b.) the implication
graphs in elle/consistency-model help cut this down.

The consistency-model graphs are also used to infer what kind of
anomalies to look for. rw-register and list-append tests take a
:consistency-models key, as well as an :anomalies key, and use them to
infer what kind of anomalies would constitute a checker error.

We now infer additional dependency graphs based on the targeted
anomalies and consistency models; you don't need to specify them by
hand.

We now report which consistency models a history is not even if it
doesn't mean a test failure. In the future, we might report anomaly
types found too?

Repeatable-read is a bit of a mess still. Got to figure out what to do
about that.