consistency-models: strong serializable, lost update
Several polishes to consistency models. We now use strong serializable
as an alias for strict serializable, which brings it in line with strong
snapshot isolation, strong session, etc.

We now have proper rules for lost update: it's prohibited by ROLA.

G1c-process no longer implies G1-process; that's not a thing.

lost-update is a special case of read skew, which is a kind of G2.

causal-cerone now implies ROLA, not just read-atomic

repeatable-read and cursor-stability, following Bailis, prohibit lost-update