list-append/merge-orders is now gone
We tried to merge together orders intelligently, but this was a.)
expensive, and b.) *already* incomplete/wrong when we had duplicated
values or inconsistent orders. In practice, whenever you see either of
those happen you already know the order inference is going to be
bonkers, so there's not much point in doing it in a complicated way. Now
we just take the longest observed order and be done with it. :-)