Add g-nonadjacent process and realtime anomaly detection
We formerly detected g-nonadjacent but did not detect realtime and
process variants, due partly to my own uncertainty about how legitimate
g-nonadjacent actually was. This actually caused someone to report an
issue in Elle where it failed to find what they considered a
strong-session-SI violation which was g-nonadjacent-process! We now
detect these anomalies. :-)