list-append: add future-read anomaly type
A future read occurs when a transaction reads a value it will write in
the future. This is an internal anomaly in the sense that it manifests
in a single transaction, but under Cerone et al it's it's actually a
violation of EXT, not INT. In Adya terms it's a singleton cycle of G1c:
T1 -wr-> T1.