Txn: refactor cycle-search SCC partitioning approach
We used to compute filtered graphs from the entire graph, then take each
SCC and look for cycles in the whole graphs using just that SCC's
transactions. This meant we spent time materializing filtered graphs
that would never actually be used. This pulls the graph selection step
all the way up: we find SCCs in the top-level graph, then for each SCC
in parallel, compute small, filtered graphs for just those vertices and
perform the cycle searches.

SCCs can never intersect, so there is no duplicated work. This also gets
rid of the extra SCC arities for the various cycle-search routines in
graph: the graph IS the SCC. I'm also hopeful this gives improved
parallelism.