Add cycle-exists anomalies
This adds a new, more precise, less informative, and more efficient
phase to cycle detection in elle.txn. In addition to performing a series
of BFS searches for specific exemplars of cycles, Elle now also computes
a series of graphs which must be acyclic under certain consistency
models. For instance, the graph:

(WW U WR U RT); RW?

(Where ; denotes sequential composition and RW? denotes RW union the
universal identity graph)

... must be acyclic under strong session snapshot isolation. If Elle can
find a strongly connected component in this graph, Elle returns an
anomaly of type :strong-session-snapshot-isolation-cycle-exists. These
existence proofs indicate the size of the smallest SCC found, and
provide the indices of the transactions in that SCC, but do not outline
a specific cycle. This is a good complement to the BFS search, which is
expensive and may fail to detect some cycles for various reasons.

This adds some optimizations which may reduce runtime in some checks,
but in our benchmark workloads, performance falls by approximately 20%
by adding this pass. We think it's still worth it. It catches several
anomalies our test suite missed, and allows precise claims to be made in
linear, as opposed to exponential time.