Represent predicate edges using distinct rels
We used to represent (e.g.) a predicate rw edge using :rw; we
distinguished between G2-item and G2 on the basis of a :predicate? field
in the explainer's linkages once a cycle was found. We now build graphs
with separate wr and wrp, rw and rwp edges, etc.

We introduce new anomaly types :G-single-item and G-nonadjacent-item,
which are subtypes of G2-item and G-single/G-nonadjacent. These
anomalies allow us to detect violations of repeatable read which are
G-single-item but not G-single (with predicates).

We redefine our cycle anomalies more precisely using these new predicate
rels. We can now do cycle existence checking precisely for repeatable
read: it must be acyclic in [union ww wwp wr wrp rw].

We extend BFSPath so that its wanted rels are a *set* of rels, any of
which may match. This lets us express a cycle which must have 'at least
one rw or rwp, and also at least one realtime'.