Remove NamedGraph and RelGraph
RelGraph allowed constant-time union and projection once we'd built the
graphs for individual rels, but we paid a huge cost in materializing the
vertices and out sets when doing Tarjan. Our new architecture makes two
costs cheaper:

1. Operations on rel sets are now so cheap as to be invisible, thanks to
   BitRels.
2. We only project SCCs to sub-relationships, we do that projection in
   parallel, and we do it less often (thanks to ruling out some
   anomalies when we find/fail-to-find cycle existence or anomalies)

We can now remove the named graphs and rel graph altogether, and obtain
a modest ~(10%?) speedup. We're hitting 50-55Khz in the sloppy test now.

We can also drop the materialize phase, since the full graph is
materialized directly.